+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     testEnvironment: 'node',
     transform: {
-        '^.+\\.ts$': ['esbuild-jest', { sourcemap: true }],
+        '^.+\\.ts$': ['esbuild-jest', { sourcemap: true, target: 'es2020' }],
     },
     testMatch: [
         '**/packages/**/*.spec.ts',