+++ packages/core/src/codecs/vec.ts
@@ -1,24 +1,23 @@
 import { concatUint8Arrays } from '@scale-codec/util';
-import JSBI from 'jsbi';
 import { Decode, DecodeResult, Encode } from '../types';
-import { encodeCompact, decodeCompact } from '../compact';
+import { encodeCompact, decodeCompact } from '../codecs/compact';
 import { decodeArray, encodeArray } from './array';
 
 export function decodeVec<T>(bytes: Uint8Array, itemDecoder: Decode<T>): DecodeResult<T[]> {
     const [length, offset] = decodeCompact(bytes);
-    const [items, itemsBytes] = decodeArray(bytes.subarray(offset), itemDecoder, JSBI.toNumber(length));
+    const [items, itemsBytes] = decodeArray(bytes.subarray(offset), itemDecoder, Number(length));
     return [items, itemsBytes + offset];
 }
 
 export function encodeVec<T>(items: T[], itemEncoder: Encode<T>): Uint8Array {
-    return concatUint8Arrays([encodeCompact(JSBI.BigInt(items.length)), encodeArray(items, itemEncoder, items.length)]);
+    return concatUint8Arrays([encodeCompact(BigInt(items.length)), encodeArray(items, itemEncoder, items.length)]);
 }
 
 /**
  * Encode `Vec<u8>` directly from the native `Uint8Array`
  */
 export function encodeUint8Vec(vec: Uint8Array): Uint8Array {
-    return concatUint8Arrays([encodeCompact(JSBI.BigInt(vec.length)), vec]);
+    return concatUint8Arrays([encodeCompact(BigInt(vec.length)), vec]);
 }
 
 /**
@@ -26,6 +25,6 @@ export function encodeUint8Vec(vec: Uint8Array): Uint8Array {
  */
 export function decodeUint8Vec(bytes: Uint8Array): DecodeResult<Uint8Array> {
     const [lenBN, offset] = decodeCompact(bytes);
-    const len = JSBI.toNumber(lenBN);
+    const len = Number(lenBN);
     return [bytes.slice(offset, offset + len), offset + len];
 }