+++ packages/docs/root/guide/components/CompactEncode.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { encodeCompact, JSBI } from '@scale-codec/core';
+import { encodeCompact } from '@scale-codec/core';
 import { hexifyBytes } from '@scale-codec/util';
 import { computed } from 'vue';
 import objectInspect from 'object-inspect';
@@ -11,7 +11,7 @@ const props = defineProps<{
 
 const result = computed(() => {
     try {
-        const encoded = encodeCompact(JSBI.BigInt(props.num));
+        const encoded = encodeCompact(BigInt(props.num));
         return props.hex ? hexifyBytes(encoded) : objectInspect(encoded);
     } catch (err) {
         return err;